Make group settings form easier to understand.
Moving from checkboxes to radio buttons means that users can choose the
setup they wish to use, rather than trying to understand the
behind-the-scenes logic. This also makes it possible to save a
meaningful “no” if you don’t want to include activity but the global
setting is “yes”. Requires refactoring the form and the save method.