Add 'mygroups' context to hgbp_group_has_children().
In an effort to handle the various kinds of directories, this changeset
expands the ‘context’ idea to include the following:
• ’normal’ - include all public and private groups, include hidden only
if the user is a member.
• ‘exclude_hidden’ - don’t include any hidden groups, regardless of
membership.
• ‘my groups’ - include only groups the user is a member of, regardless
of status.

I added a bunch of test cases that test the various combinations, of
which there are tons, given group status, user status and context.