Remove hgbp_get_child_group_ids().
`hgbp_get_child_group_ids()` was added to make use of caching to store
child group IDs. However, because the underlying caching in
`BP_Groups_Group::get()` is so good, I removed my on-the-top caching.
(Thanks to @boonebgorges.) This function is also not needed as a result
of the change.