Change logic to guess when to apply hierarchy to main groups directory
At its best, this is a guess. But, I think it’s better to be on the
conservative side here, and apply the hierarchical filter (showing
top-level groups only) when the user isn’t hunting for a specific
group. For instance, if you search for a group that is a second-level
group, and the parent group isn’t in the results set, you’d never be
able to reach that second level group. Awkward. So, when in doubt,
don’t force the groups loop to display only top level groups.