Improve parent group selection interaction.
Well, this is more complicated than it seems. Because the list of
possible parent groups is dependent on which user is making the
selection, it’s important to make it easy to know what your choice is
doing.
• Add an info paragraph naming the current parent group or that it’s
hidden and you can’t see it.
• Adding an explicit “Keep current parent group” option which is
helpful if you aren’t able to choose the current parent group (that
someone else was allowed to).
• Improve layout/styling of the form.