Avoid conflicts when checking action variables for child group slugs.
Introduce `hgbp_child_group_exists()` which is like `groups_get_id()`
but limited to children of a specific group. Avoids slug collisions
between group tab names and groups with the same slug.  For instance,
if there's a unrelated group called "Docs", you want the "docs" tab of
a group to ignore that group and return the docs pane for the current
group.
Caveat: If you create a child group with the same slug as a tab of the
parent
group, you'll always get the child group.