Optimize queries run when registering group extension.
The BP Group Extension is instantiated on every page load, and when
we’re not in a single group, the query uses the `parent_id` of 0,
resulting in an expensive query. So, don’t try to resolve the single
group tab name and who can access it unless we have a non-zero group ID
to work with.